fix(runtime): Fix Runtime component adding in vue2
Within Vue2, the addComponent will occasionally fail due to the component / script signature being inconsistent. To fix this we're using the vuetify addComponent code as it's known to work

Vue3 does not seem to have this issue so it's still using the old version

Have also taken the opportunity to update the dependencies so the build does not include them.